extract requestability resolver
