+++ lib/response_massager.js
@@ -1,8 +1,9 @@
 const LocationLabelUpdater = require('./location_label_updater')
 const AvailabilityResolver = require('./availability_resolver.js')
 const parallelFieldsExtractor = require('./parallel-fields-extractor')
-const {isAeonUrl} = require('../lib/util')
+const { isAeonUrl } = require('../lib/util')
 const FulfillmentResolver = require('./fulfillment_resolver')
+const RequestabilityResolver = require('./requestability_resolver')
 // const addNumItemsMatched = require('./item-match-numerator')
 
 class ResponseMassager {
@@ -67,7 +68,7 @@ class ResponseMassager {
 
     // Update ES response with updated availability from SCSB:
     const updatedWithAvailability = (new AvailabilityResolver(response))
-      .responseWithUpdatedAvailability(request, options)
+      .responseWithUpdatedAvailability(options)
 
     // Update ES response with NYPL-Core labels:
     return updatedWithAvailability
@@ -76,6 +77,8 @@ class ResponseMassager {
           .responseWithUpdatedLabels()
       })
       .then((response) => new FulfillmentResolver(response).responseWithFulfillment())
+      .then((response) => new RequestabilityResolver(response)
+        .fixItemRequestability(response, request))
   }
 }
