+++ test/availability_resolver.test.js
@@ -1,17 +1,14 @@
 const { expect } = require('chai')
 const sinon = require('sinon')
 
+const RequestabilityResolver = require('../lib/requestability_resolver')
 const AvailabilityResolver = require('../lib/availability_resolver.js')
 const elasticSearchResponse = require('./fixtures/elastic_search_response.js')
-const eddElasticSearchResponse = require('./fixtures/edd_elastic_search_response')
-const specRequestableElasticSearchResponse = require('./fixtures/specRequestable-es-response')
 const recapScsbQueryMatch = require('./fixtures/recap-scsb-query-match')
 const recapScsbQueryMismatch = require('./fixtures/recap-scsb-query-mismatch')
 const logger = require('../lib/logger')
 const scsbClient = require('../lib/scsb-client')
 const esClient = require('../lib/es-client')
-const DeliveryLocationsResolver = require('../lib/delivery-locations-resolver.js')
-const requestabilityDetermination = require('../lib/requestability_determination')
 
 const itemAvailabilityResponse = [
   {
@@ -141,20 +138,6 @@ describe('Response with updated availability', function () {
       })
   })
 
-  it('will set requestable to false for an item not found in ReCAP', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    let indexedButNotAvailableInSCSBURI = 'i22566485'
-
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((modifiedResponse) => {
-        // Find the modified item in the response:
-        let theItem = modifiedResponse.hits.hits[0]._source.items.find((item) => item.uri === indexedButNotAvailableInSCSBURI)
-        // Our fakeRESTClient said its barcode doesn't exist, so it should appear with `requestable` false
-        expect(theItem.requestable[0]).to.equal(false)
-      })
-  })
-
   it('includes the latest availability status of items', function () {
     let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
 
@@ -177,104 +160,6 @@ describe('Response with updated availability', function () {
         })
         expect(availableItem.status[0].id).to.equal('status:a')
         expect(availableItem.status[0].label).to.equal('Available')
-      })
-  })
-
-  it('marks ReCAP items that are in requestable locations and have delivery locations as physRequestable', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((response) => {
-        var items = response.hits.hits[0]._source.items
-
-        var availableItem = items.find((item) => {
-          return item.uri === 'i10283664'
-        })
-        expect(availableItem.physRequestable).to.equal(true)
-        expect(availableItem.eddRequestable).to.equal(true)
-      })
-  })
-
-  it('marks ReCAP items that are in unrequestable locations as not eddRequestable nor physRequestable', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((response) => {
-        var items = response.hits.hits[0]._source.items
-
-        var availableItem = items.find((item) => {
-          return item.uri === 'i102836649-unrequestable'
-        })
-        expect(availableItem.physRequestable).to.equal(false)
-        expect(availableItem.eddRequestable).to.equal(false)
-      })
-  })
-
-  it('marks SCSB Available items (that are indexed as Not Available) as requestable', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((response) => {
-        var items = response.hits.hits[0]._source.items
-
-        // A ReCAP item with Discovery status 'Not Available', but SCSB
-        // status 'Available' should be made requestable:
-        var availableItem = items.find((item) => {
-          return item.uri === 'i10283664'
-        })
-        expect(availableItem.requestable[0]).to.equal(true)
-      })
-  })
-
-  it('marks SCSB Not-Available items as requestable', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((response) => {
-        var items = response.hits.hits[0]._source.items
-
-        // A ReCAP item with SCSB status 'Not Available' should be made not
-        // requestable:
-        var notAvailableItem = items.find((item) => item.uri === 'i102836649')
-        expect(notAvailableItem.requestable[0]).to.equal(true)
-        expect(notAvailableItem.status[0].label).to.equal('Not available')
-      })
-  })
-
-  it('marks on-site (loc:scff2) Available items as requestable', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((response) => {
-        var items = response.hits.hits[0]._source.items
-
-        var availableItem = items.find((item) => item.uri === 'i10283665')
-        expect(availableItem.requestable[0]).to.equal(true)
-      })
-  })
-
-  it('marks on-site (loc:scff2) Not-Available items as not requestable', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((response) => {
-        var items = response.hits.hits[0]._source.items
-
-        var notAvailableItem = items.find((item) => item.uri === 'i10283665777')
-        expect(notAvailableItem.requestable[0]).to.equal(false)
-      })
-  })
-
-  it('marks on-site (loc:scff2) Available items as not requestable if "no-on-site-edd" feature flag is set', function () {
-    let availabilityResolver = new AvailabilityResolver(elasticSearchResponse.fakeElasticSearchResponseNyplItem())
-
-    process.env.FEATURES = 'no-on-site-edd'
-    return availabilityResolver.responseWithUpdatedAvailability()
-      .then((response) => {
-        var items = response.hits.hits[0]._source.items
-
-        var availableItem = items.find((item) => item.uri === 'i10283665')
-        expect(availableItem.eddRequestable).to.equal(false)
       })
   })
 
@@ -307,66 +192,12 @@ describe('Response with updated availability', function () {
     })
   })
 
-  describe('Special collections items', function () {
-    let availabilityResolver = null
-    before(function () {
-      availabilityResolver = new AvailabilityResolver(specRequestableElasticSearchResponse())
-    })
-    it('marks items as specRequestable when there is an aeonURL present', function () {
-      return availabilityResolver.responseWithUpdatedAvailability()
-        .then((response) => {
-          const items = response.hits.hits[0]._source.items
-          const specRequestableItem = items.find((item) => item.uri === 'i22566485')
-          expect(specRequestableItem.specRequestable).to.equal(true)
-        })
-    })
-    it('marks items as not specRequestable when there is no aeonURL present', function () {
-      return availabilityResolver.responseWithUpdatedAvailability()
-        .then((response) => {
-          const items = response.hits.hits[0]._source.items
-          const specRequestableItem = items.find((item) => item.uri === 'i10283665')
-          expect(specRequestableItem.specRequestable).to.equal(false)
-        })
-    })
-  })
-
-  describe('eddRequestable items', function () {
-    let availabilityResolver = null
-    before(function () {
-      availabilityResolver = new AvailabilityResolver(eddElasticSearchResponse())
-    })
-    it('marks items eddRequestable:true when its reCAP code is listed as such in nypl-core', () => {
-      return availabilityResolver.responseWithUpdatedAvailability()
-        .then((response) => {
-          var items = response.hits.hits[0]._source.items
-
-          // A ReCAP item with customer code NA (eddRequestable = true)
-          var eddItem = items.find((item) => {
-            return item.uri === 'i102836649'
-          })
-          expect(eddItem.eddRequestable).to.equal(true)
-        })
-    })
-    it('marks items eddRequestable:false when its reCAP code is listed as such in nypl-core', () => {
-      return availabilityResolver.responseWithUpdatedAvailability()
-        .then((response) => {
-          var items = response.hits.hits[0]._source.items
-
-          // A ReCAP item with customer code NC (eddRequestable = false)
-          var nonEddItem = items.find((item) => {
-            return item.uri === 'i10283664'
-          })
-          expect(nonEddItem.eddRequestable).to.equal(false)
-        })
-    })
-  })
-
   describe('checks recapCustomerCodes when options specifies', () => {
     let availabilityResolver = null
     it('logs an error when item\'s code does not match SCSB', () => {
       availabilityResolver = new AvailabilityResolver(recapScsbQueryMismatch())
       const loggerSpy = sinon.spy(logger, 'error')
-      return availabilityResolver.responseWithUpdatedAvailability(null, { queryRecapCustomerCode: true })
+      return availabilityResolver.responseWithUpdatedAvailability({ queryRecapCustomerCode: true })
         .then(() => {
           expect(loggerSpy.calledOnce).to.equal(true)
           logger.error.restore()
@@ -376,7 +207,8 @@ describe('Response with updated availability', function () {
     it('updates recapCustomerCode when item\'s code does not match SCSB', () => {
       return availabilityResolver.responseWithUpdatedAvailability()
         .then((response) => {
-          let items = response.hits.hits[0]._source.items
+          let responseWithRequestability = new RequestabilityResolver(response).fixItemRequestability()
+          let items = responseWithRequestability.hits.hits[0]._source.items
           // A ReCAP item with customer code XX
           const queryItem = items.find((item) => {
             return item.uri === 'i10283667'
@@ -499,7 +331,7 @@ describe('Response with updated availability', function () {
       const recapBarcodesByStatus = {
         Available: ['barcode1', 'barcode2']
       }
-      return availabilityResolver.responseWithUpdatedAvailability({}, { recapBarcodesByStatus })
+      return availabilityResolver.responseWithUpdatedAvailability({ recapBarcodesByStatus })
         .then((modifiedResponse) => {
           const buckets = modifiedResponse.aggregations.item_status._nested.buckets
           expect(buckets).to.deep.equal([
@@ -517,7 +349,7 @@ describe('Response with updated availability', function () {
         Available: ['barcode1'],
         'Not Available': ['barcode2']
       }
-      return availabilityResolver.responseWithUpdatedAvailability({}, { recapBarcodesByStatus })
+      return availabilityResolver.responseWithUpdatedAvailability({ recapBarcodesByStatus })
         .then((modifiedResponse) => {
           const buckets = modifiedResponse.aggregations.item_status._nested.buckets
           expect(buckets).to.deep.equal([
@@ -525,51 +357,6 @@ describe('Response with updated availability', function () {
             { key: 'status:na||Not available', doc_count: 1 }
           ])
         })
-    })
-  })
-  describe('_fixPhysRequestablility', () => {
-    let deliveryResolverStub
-    let holdingLocationStub
-    beforeEach(() => {
-      deliveryResolverStub = sinon.stub(DeliveryLocationsResolver, 'deliveryLocationsByM2CustomerCode')
-      holdingLocationStub = sinon.stub(requestabilityDetermination, 'requestableBasedOnHoldingLocation')
-    })
-    afterEach(() => {
-      holdingLocationStub.restore()
-      deliveryResolverStub.restore()
-      process.env.ROMCOM_MAX_XA_BNUM = null
-    })
-
-    it('returns true for item with requestable m2 customer code', () => {
-      const item = { m2CustomerCode: ['vk'] }
-      deliveryResolverStub.returns(['loc:ma82, loc:456'])
-      holdingLocationStub.returns(true)
-      expect(AvailabilityResolver.prototype._fixPhysRequestability(item, false)).to.equal(true)
-    })
-    it('returns false for item with no m2 customer code', () => {
-      const item = { notM2: ['ab'] }
-      holdingLocationStub.returns(false)
-      expect(AvailabilityResolver.prototype._fixPhysRequestability(item, false)).to.equal(false)
-      expect(deliveryResolverStub.called).to.equal(false)
-    })
-    it('returns false for item with m2 customer code that has no delivery locations returned', () => {
-      const item = { m2CustomerCode: ['vk'] }
-      deliveryResolverStub.returns([])
-      holdingLocationStub.returns(true)
-      expect(AvailabilityResolver.prototype._fixPhysRequestability(item, false)).to.equal(false)
-    })
-    it('overrides physRequestability for XA bnumbers when env variable is set', () => {
-      process.env.ROMCOM_MAX_XA_BNUM = 'b000000'
-      const item = { m2CustomerCode: ['XA'] }
-      deliveryResolverStub.returns(['loc:ma82, loc:456'])
-      holdingLocationStub.returns(true)
-      expect(AvailabilityResolver.prototype._fixPhysRequestability(item, false, 'b000001')).to.equal(false)
-    })
-    it('does not override physRequestability for XA bnumbers when env variable is not set', () => {
-      const item = { m2CustomerCode: ['XA'] }
-      deliveryResolverStub.returns(['loc:ma82, loc:456'])
-      holdingLocationStub.returns(true)
-      expect(AvailabilityResolver.prototype._fixPhysRequestability(item, false, 'b000001')).to.equal(true)
     })
   })
 })
